requirements.txt
albumentations>=1.3.0
opencv-python>=4.8.0
tqdm>=4.66.0
numpy>=1.24.0

# For diffusion model training
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
accelerate>=0.20.0
Pillow>=10.0.0

